realize factory for handlers
