+++ softgroup/basis/common-dao/common-dao-api/src/main/java/Conversation.java
@@ -1,19 +1,16 @@
-package com.softgroup.common.dao.api.entities;
-
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Id;
 import javax.persistence.Table;
-import java.io.Serializable;
 
 /**
- * Created by User on 21.03.2017.
+ * Created by Javier on 06.04.2017.
  */
 
 
 @Entity
 @Table(name = "conversation")
-public class ConversationEntity implements Serializable {
+public class Conversation  { // FIXME: 06.04.2017 RENAME CLASS
 
     @Id
     @Column(name = "id")
@@ -21,6 +18,9 @@ public class ConversationEntity implements Serializable {
 
     @Column(name = "name")
     private String name;
+
+    @Column(name = "admin_id")
+    private String adminId;
 
     @Column(name = "logo_image_uri")
     private String logoImageUri;
@@ -30,6 +30,16 @@ public class ConversationEntity implements Serializable {
 
     @Column(name = "last_message_index")
     private long lastMessageIndex;
+
+
+
+    public String getAdminId() {
+        return adminId;
+    }
+
+    public void setAdminId(String adminId) {
+        this.adminId = adminId;
+    }
 
     public String getId() {
         return id;
@@ -70,4 +80,5 @@ public class ConversationEntity implements Serializable {
     public void setLastMessageIndex(long lastMessageIndex) {
         this.lastMessageIndex = lastMessageIndex;
     }
+
 }
