--- softgroup/basis/common-dao/common-dao-api/src/main/java/com/softgroup/common/dao/api/entities/MessageEntity.java
@@ -1,100 +1,0 @@
-package com.softgroup.common.dao.api.entities;
-
-import javax.persistence.Column;
-import javax.persistence.Id;
-import java.io.Serializable;
-
-/**
- * Created by User on 21.03.2017.
- */
-public class MessageEntity implements Serializable {
-
-    @Id
-    @Column (name = "id")
-    private String id;
-
-    @Column (name = "sender_id")
-    private String senderId;
-
-    @Column(name = "message_type")
-    private Integer messageType;
-
-    @Column(name = "payload")
-    private  String payLoad;
-
-    @Column(name = "status")
-    private Integer status;
-
-    @Column(name = "create_time")
-    private Long createTime;
-
-    @Column (name = "server_receive_time")
-    private Long serverReceiveTime;
-
-    @Column(name = "message_index")
-    private Long messageIndex;
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getSenderId() {
-        return senderId;
-    }
-
-    public void setSenderId(String senderId) {
-        this.senderId = senderId;
-    }
-
-    public Integer getMessageType() {
-        return messageType;
-    }
-
-    public void setMessageType(Integer messageType) {
-        this.messageType = messageType;
-    }
-
-    public String getPayLoad() {
-        return payLoad;
-    }
-
-    public void setPayLoad(String payLoad) {
-        this.payLoad = payLoad;
-    }
-
-    public Integer getStatus() {
-        return status;
-    }
-
-    public void setStatus(Integer status) {
-        this.status = status;
-    }
-
-    public Long getCreateTime() {
-        return createTime;
-    }
-
-    public void setCreateTime(Long createTime) {
-        this.createTime = createTime;
-    }
-
-    public Long getServerReceiveTime() {
-        return serverReceiveTime;
-    }
-
-    public void setServerReceiveTime(Long serverReceiveTime) {
-        this.serverReceiveTime = serverReceiveTime;
-    }
-
-    public Long getMessageIndex() {
-        return messageIndex;
-    }
-
-    public void setMessageIndex(Long messageIndex) {
-        this.messageIndex = messageIndex;
-    }
-}
