--- softgroup/basis/common-dao/common-dao-api/src/main/java/com/softgroup/common/dao/api/entities/ProfileStatusEntity.java
@@ -1,7 +1,0 @@
-package com.softgroup.common.dao.api.entities;
-
-/**
- * Created by User on 21.03.2017.
- */
-public class ProfileStatusEntity {
-}
