--- softgroup/basis/common-dao/common-dao-api/src/main/java/com/softgroup/common/dao/api/entities/ProfileStrictEntity.java
@@ -1,22 +1,0 @@
-package com.softgroup.common.dao.api.entities;
-
-import java.io.Serializable;
-
-/**
- * Created by User on 21.03.2017.
- */
-
-
-
-public class ProfileStrictEntity implements Serializable
-
-{
-
-
-        }
-
-
-/*id	String	ключ
-phone_number	String	полный номер телефона
-name	String	имя пользователя*/
-
